requirements.txt
torch
torchvision
Pillow
matplotlib
tqdm
python-dotenv
discord
numpy